Update and rename readme to Area.py
Create a program named area.py. The program must:
 Ask the user: The measurement unit (inches, feet, centimeters, etc.)
 The length of the rectangle in the units specified (e.g., “What is the length in inches?”) Float values must be accepted.
 The width of the rectangle in the units
specified (e.g., “What is the length in centimeters?”). Float values must be accepted.
 Calculate the area of the rectangle (defined as length * width).
 Display the area as a full statement rounded
to two decimal places (e.g., “The area of the rectangle is 12.70 square centimeters.”)
 NOTE: the values in italics at the end of lines 2-4 in the example are NOT hard coded. They are user input & could be different. Likewise, in the 3 print statements, the word ”meters" and the number 9.00 are NOT hard coded. They are filled in from variable names & will change depending on user input.

Welcome to the Rectangle Area Calculator!
What is your measurement unit (in., ft., cm., etc.)? meters
What is the length of the rectangle in meters? 2.5
What is the width of the rectangle in meters? 3.6
Your rectangle is 9.00 square meters.