Final_project.py
When adding a pizza, the program
needs to display the number of the
pizza being ordered.
 Your program must download the
ingredients.json file from the
internet! Do not download it yourself
and save it with your program.
 The URL is https://itec-
minneapolis.s3.us-west-
2.amazonaws.com/ingredients.json
 All the options for the pizza must come
from the file downloaded from the
internet—there must be no categories,
choices, or toppings hard-coded in
your program!
'base_options' is a list of options that make up the base of the
pizza (crust, sauce, cheese, etc.).
 Each base_option has a category and several options within it.
 Your program will loop through each of the base options and
ask the user to select the category by giving the options as a
menu
toppings' is simpler than base options.
 Each topping will be a simple yes/no
question.
 Your program will loop through each of
the toppings and ask the user if they
want the topping.
Viewing the order must display each pizza
(numbered), along with all options and
toppings for that pizza.
 Each option and topping must have its
price displayed.
 The total for each individual pizza must be
displayed.
 The grand total for the whole order must be
displayed.
 All prices must be in a right-aligned column
with a dollar sign and decimal point.
When placing the order, it must be
displayed as in Viewing the Order.
 The full order must be written to a file
in JSON format—this is to simulate
sending the order to the store over the
internet.
 A message should be shown that the
order was placed, and then the
program should exit.