Update and rename read me to Book_List.py
The starter file collects book prices into a list, and totals,
averages & displays the info. The revised program must:
 Add a feature to collect the book titles
 Store the book titles into their own list, parallel with prices
 Display the titles with the prices as shown on the following slide

Lab Section 2: Book List, Cont. This is the output from the starter program. You must modify the starter program to produce this output. Note the prompts to the user for the titles, and the titles in the price list.