Update and rename readme to Float_Validation.py
Prompt the user for a number with or without a decimal point or minus symbol.
 Use a regular expression to check if the input is a valid number:
 The minus sign is optional and must be at the beginning present.
 There must be one or more digits before the decimal point, if it is present.
 The decimal point is optional and only one is allowed.
 There must be one or more digits after the decimal point.
 No characters other than digits (0-9), the decimal and a minus sign may be present.
 Output a message to the user telling them whether their input is a valid number. Include the number in the message if it is valid.
Section 2: Examples Valid Input
Enter a number. Negatives and decimals are allowed:
3.14159265
3.14159265 is a valid number!
Enter a number. Negatives and decimals are allowed:
-123.456
-123.456 is a valid number!
Enter a number. Negatives and decimals are allowed:
-0.99
-0.99 is a valid number!

Invalid Input
Enter a number. Negatives and decimals are allowed:
-.25
This does not look like a valid number.
Enter a number. Negatives and decimals are allowed:
1.0.1
This does not look like a valid number.
Enter a number. Negatives and decimals are allowed:
0xAE
This does not look like a valid number